refactor: 💡 submit move to hooks file
